fix(partners-list) use atr for dropdown selection
Previously text-content was used to bind to the selection
value for fields like partner-type. This is less flexible
and also created problems in firefox/IE with empty text-nodes.
Adding name attributes to the dropdowns and selecting based on that
fixes this.

Also small style update to the list in page-partners. List should not
have the same layout that partners-details and overview do. Now it's
centered on the page.